vaxis: use request/response model for OSC 11
Use a request/response model for the OSC 11 implementation to make it
concurrency safe and avoid data races (similar to the OSC 99
implementation). Remove the chBg channel and the mutex in the writer
since they are no longer needed.

If the terminal requests the background color, the OSC 11 request will
be executed during the rendering loop. The OSC 11 response is then
processed in vx.handleSequence() which sends a BgColor event to the main
event loop.